+++ src/bin/profparser.rs
@@ -365,7 +365,7 @@ impl MergeCommand {
             !self.input.is_empty(),
             "No input files selected. See merge --help"
         );
-        let profile = merge_profiles(&self.input)?;
+        let _profile = merge_profiles(&self.input)?;
         // Now to write it out?
         Ok(())
     }