+++ src/instrumentation_profile/types.rs
@@ -55,7 +55,6 @@ impl Symtab {
     /// otherwise.
     pub fn add_func_name(&mut self, name: String, endianness: Option<Endianness>) {
         let hash = match endianness {
-            Some(Endianness::Little) => compute_hash(&name),
             Some(Endianness::Big) => compute_be_hash(&name),
             _ => compute_hash(&name),
         };
@@ -154,17 +153,26 @@ impl InstrumentationProfile {
     }
 
     pub fn merge_record(&mut self, record: &NamedInstrProfRecord) {
-        if self.symtab.contains(record.hash_unchecked()) {
-            // Find the record and merge things. 0 hashed records should have no counters in the
-            // code and otherwise we'll ignore the change that truncated md5 hashes can collide
-            if let Some(rec) = self.records.iter_mut().find(|x| x.name == record.name) {
-                rec.record.merge(&record.record);
+        if let Some(name) = record.name.as_ref() {
+            let hash = compute_hash(name);
+            if self.symtab.contains(hash) {
+                // Find the record and merge things. 0 hashed records should have no counters in the
+                // code and otherwise we'll ignore the change that truncated md5 hashes can collide
+                if let Some(rec) = self.records.iter_mut().find(|x| x.name == record.name) {
+                    rec.record.merge(&record.record);
+                }
+            } else {
+                if let Some(alt_hash) = record.hash {
+                    if self.symtab.contains(alt_hash) {
+                        if let Some(rec) = self.records.iter_mut().find(|x| x.name == record.name) {
+                            rec.record.merge(&record.record);
+                        }
+                    }
+                } else {
+                    self.symtab.names.insert(hash, record.name_unchecked());
+                    self.records.push(record.clone());
+                }
             }
-        } else {
-            self.symtab
-                .names
-                .insert(record.hash_unchecked(), record.name_unchecked());
-            self.records.push(record.clone());
         }
     }
 
@@ -320,7 +328,7 @@ pub struct ValueProfRecord {
 }
 
 impl ValueProfData {
-    fn deserialize_to(&self, record: &mut InstrProfRecord, symtab: Option<&Symtab>) {
+    fn deserialize_to(&self, _record: &mut InstrProfRecord, _symtab: Option<&Symtab>) {
         if self.num_value_kinds == 0 {
             return;
         }