+++ tests/profdata.rs
@@ -195,7 +195,7 @@ fn show_profdatas() {
 }
 
 #[test]
-#[ignore]
+#[cfg_attr(llvm_15, ignore)]
 fn merge() {
     let data = get_data_dir();
     let files = [
@@ -205,6 +205,19 @@ fn merge() {
     ];
 
     check_merge_command(&files, "foo_results");
+}
+
+#[test]
+fn profraw_merging() {
+    let premerge_1 = data_root_dir().join("misc").join("premerge_1.profraw");
+    let premerge_2 = data_root_dir().join("misc").join("premerge_2.profraw");
+    let merged = data_root_dir().join("misc").join("merged.profdata");
+
+    let expected_merged = merge_profiles(&[merged]).unwrap();
+    let merged = merge_profiles(&[premerge_1, premerge_2]).unwrap();
+
+    assert_eq!(merged.symtab, expected_merged.symtab);
+    assert_eq!(merged.records, expected_merged.records);
 }
 
 #[test]